Port to Python 3
Mostly a trivial port but some complications.

int/int in Python 3 returns a float, not an int.  For the purposes of this
script we don't care about that level of accuracy so force the result to an
int().

Python 3's urllib doesn't like URLs containing embedded username/password (tries
to parse the password as a port number), so add an Authorization header
explicitly.